translator/requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0
lxml>=4.9.0
